minor bugfix

+++ src/ca/mcgill/ecse223/block/application/Block223Application.java
@@ -31,7 +31,7 @@ public class Block223Application {
  		return block223;
 	}
 	public static Block223 resetBlock223() {
-		block223.delete();
+		if(block223!=null)block223.delete();
 		Block223Application.setCurrentGame(null);
 		Block223Application.setCurrentPlayableGame(null);
 		block223 = Block223Persistence.load();
